Moved GDB packet IO thread to separate module.
- Replaced loggers created within the classes with module logger.
