+++ pyocd/gdbserver/gdbserver.py
@@ -14,6 +14,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import logging
+import threading
+from struct import unpack
+from time import (sleep, time)
+import sys
+import six
+from xml.etree.ElementTree import (Element, SubElement, tostring)
+
 from ..core import exceptions
 from ..core.target import Target
 from ..flash.loader import (FlashLoader, FlashEraser)
@@ -31,29 +39,17 @@ from .context_facade import GDBDebugContextFacade
 from .symbols import GDBSymbolProvider
 from ..rtos import RTOS
 from . import signals
-import logging, threading, socket
-from struct import unpack
-from time import (sleep, time)
-import sys
-import six
-from six.moves import queue
-from xml.etree.ElementTree import (Element, SubElement, tostring)
-
-CTRL_C = b'\x03'
+from .packet_io import (
+    CTRL_C,
+    checksum,
+    ConnectionClosedException,
+    GDBServerPacketIOThread,
+    )
 
 LOG = logging.getLogger(__name__)
 
-TRACE_MEM = LOG.getChild("trace_mem")
+TRACE_MEM = LOG.getChild("trace.mem")
 TRACE_MEM.setLevel(logging.CRITICAL)
-
-TRACE_ACK = LOG.getChild("trace_ack")
-TRACE_ACK.setLevel(logging.CRITICAL)
-
-TRACE_PACKETS = LOG.getChild("trace_packet")
-TRACE_PACKETS.setLevel(logging.CRITICAL)
-
-def checksum(data):
-    return ("%02x" % (sum(six.iterbytes(data)) % 256)).encode()
 
 def unescape(data):
     """! @brief De-escapes binary data from Gdb.
@@ -95,172 +91,6 @@ class GDBError(exceptions.Error):
     """! @brief Error communicating with GDB."""
     pass
 
-class ConnectionClosedException(Exception):
-    """! @brief Exception used to signal the GDB server connection closed."""
-    pass
-
-class GDBServerPacketIOThread(threading.Thread):
-    """! @brief Packet I/O thread.
-    
-    This class is a thread used by the GDBServer class to perform all RSP packet I/O. It
-    handles verifying checksums, acking, and receiving Ctrl-C interrupts. There is a queue
-    for received packets. The interface to this queue is the receive() method. The send()
-    method writes outgoing packets to the socket immediately.
-    """
-    
-    def __init__(self, abstract_socket):
-        super(GDBServerPacketIOThread, self).__init__()
-        self.name = "gdb-packet-thread-port%d" % abstract_socket.port
-        self.log = LOG.getChild('gdbpacket')
-        self._abstract_socket = abstract_socket
-        self._receive_queue = queue.Queue()
-        self._shutdown_event = threading.Event()
-        self.interrupt_event = threading.Event()
-        self.send_acks = True
-        self._clear_send_acks = False
-        self._buffer = b''
-        self._expecting_ack = False
-        self.drop_reply = False
-        self._last_packet = b''
-        self._closed = False
-        self.setDaemon(True)
-        self.start()
-
-    def set_send_acks(self, ack):
-        if ack:
-            self.send_acks = True
-        else:
-            self._clear_send_acks = True
-
-    def stop(self):
-        self._shutdown_event.set()
-
-    def send(self, packet):
-        if self._closed or not packet:
-            return
-        if not self.drop_reply:
-            self._last_packet = packet
-            self._write_packet(packet)
-        else:
-            self.drop_reply = False
-            self.log.debug("GDB dropped reply %s", packet)
-
-    def receive(self, block=True):
-        if self._closed:
-            raise ConnectionClosedException()
-        while True:
-            try:
-                # If block is false, we'll get an Empty exception immediately if there
-                # are no packets in the queue. Same if block is true and it times out
-                # waiting on an empty queue.
-                return self._receive_queue.get(block, 0.1)
-            except queue.Empty:
-                # Only exit the loop if block is false or connection closed.
-                if not block:
-                    return None
-                if self._closed:
-                    raise ConnectionClosedException()
-
-    def run(self):
-        self.log.debug("Starting GDB server packet I/O thread")
-
-        self._abstract_socket.set_timeout(0.01)
-
-        while not self._shutdown_event.is_set():
-            try:
-                data = self._abstract_socket.read()
-
-                # Handle closed connection
-                if len(data) == 0:
-                    self.log.debug("GDB packet thread: other side closed connection")
-                    self._closed = True
-                    break
-
-                TRACE_PACKETS.debug('-->>>> GDB read %d bytes: %s', len(data), data)
-
-                self._buffer += data
-            except socket.error:
-                pass
-
-            if self._shutdown_event.is_set():
-                break
-
-            self._process_data()
-
-        self.log.debug("GDB packet thread stopping")
-
-    def _write_packet(self, packet):
-        TRACE_PACKETS.debug('--<<<< GDB send %d bytes: %s', len(packet), packet)
-
-        # Make sure the entire packet is sent.
-        remaining = len(packet)
-        while remaining:
-            written = self._abstract_socket.write(packet)
-            remaining -= written
-            if remaining:
-                packet = packet[written:]
-
-        if self.send_acks:
-            self._expecting_ack = True
-
-    def _check_expected_ack(self):
-        # Handle expected ack.
-        c = self._buffer[0:1]
-        if c in (b'+', b'-'):
-            self._buffer = self._buffer[1:]
-            TRACE_ACK.debug('got ack: %s', c)
-            if c == b'-':
-                # Handle nack from gdb
-                self._write_packet(self._last_packet)
-                return
-
-            # Handle disabling of acks.
-            if self._clear_send_acks:
-                self.send_acks = False
-                self._clear_send_acks = False
-        else:
-            self.log.debug("GDB: expected n/ack but got '%s'", c)
-
-    def _process_data(self):
-        """! @brief Process all incoming data until there are no more complete packets."""
-        while len(self._buffer):
-            if self._expecting_ack:
-                self._expecting_ack = False
-                self._check_expected_ack()
-
-            # Check for a ctrl-c.
-            if len(self._buffer) and self._buffer[0:1] == CTRL_C:
-                self.interrupt_event.set()
-                self._buffer = self._buffer[1:]
-
-            try:
-                # Look for complete packet and extract from buffer.
-                pkt_begin = self._buffer.index(b"$")
-                pkt_end = self._buffer.index(b"#") + 2
-                if pkt_begin >= 0 and pkt_end < len(self._buffer):
-                    pkt = self._buffer[pkt_begin:pkt_end + 1]
-                    self._buffer = self._buffer[pkt_end + 1:]
-                    self._handling_incoming_packet(pkt)
-                else:
-                    break
-            except ValueError:
-                # No complete packet received yet.
-                break
-
-    def _handling_incoming_packet(self, packet):
-        # Compute checksum
-        data, cksum = packet[1:].split(b'#')
-        computedCksum = checksum(data)
-        goodPacket = (computedCksum.lower() == cksum.lower())
-
-        if self.send_acks:
-            ack = b'+' if goodPacket else b'-'
-            self._abstract_socket.write(ack)
-            TRACE_ACK.debug(ack)
-
-        if goodPacket:
-            self._receive_queue.put(packet)
-
 class GDBServer(threading.Thread):
     """! @brief GDB remote server thread.
     
@@ -278,7 +108,6 @@ class GDBServer(threading.Thread):
             self.core = core
             self.target = self.board.target.cores[core]
         self.name = "gdb-server-core%d" % self.core
-        self.log = LOG.getChild('gdbserver')
         self.abstract_socket = None
         self.port = session.options.get('gdbserver_port')
         if self.port != 0:
@@ -336,7 +165,7 @@ class GDBServer(threading.Thread):
             console_file = self.telnet_server
             semihost_console = semihost.ConsoleIOHandler(self.telnet_server)
         else:
-            self.log.info("Semihosting will be output to console")
+            LOG.info("Semihosting will be output to console")
             console_file = sys.stdout
             self.telnet_server = None
             semihost_console = semihost_io_handler
@@ -345,7 +174,7 @@ class GDBServer(threading.Thread):
         self._swv_reader = None
         if session.options.get("enable_swv"):
             if "swv_system_clock" not in session.options:
-                self.log.warning("Cannot enable SWV due to missing swv_system_clock option")
+                LOG.warning("Cannot enable SWV due to missing swv_system_clock option")
             else:
                 sys_clock = int(session.options.get("swv_system_clock"))
                 swo_clock = int(session.options.get("swv_clock"))
@@ -410,10 +239,10 @@ class GDBServer(threading.Thread):
             self.shutdown_event.set()
             while self.isAlive():
                 pass
-            self.log.info("GDB server thread killed")
+            LOG.info("GDB server thread killed")
 
     def _cleanup(self):
-        self.log.debug("GDB server cleaning up")
+        LOG.debug("GDB server cleaning up")
         if self.packet_io:
             self.packet_io.stop()
             self.packet_io = None
@@ -435,7 +264,7 @@ class GDBServer(threading.Thread):
         self.current_thread_id = 0
 
     def run(self):
-        self.log.info('GDB server started on port %d', self.port)
+        LOG.info('GDB server started on port %d', self.port)
 
         while True:
             try:
@@ -458,13 +287,13 @@ class GDBServer(threading.Thread):
                 if self.detach_event.isSet():
                     continue
 
-                self.log.info("One client connected!")
+                LOG.info("One client connected!")
                 self._run_connection()
-                self.log.info("Client disconnected!")
+                LOG.info("Client disconnected!")
                 self._cleanup_for_next_connection()
 
             except Exception as e:
-                self.log.error("Unexpected exception: %s", e, exc_info=self.session.log_tracebacks)
+                LOG.error("Unexpected exception: %s", e, exc_info=self.session.log_tracebacks)
 
     def _run_connection(self):
         while True:
@@ -482,17 +311,17 @@ class GDBServer(threading.Thread):
                         self.is_target_running = False
                         self.send_stop_notification()
                     else:
-                        self.log.error("Got unexpected ctrl-c, ignoring")
+                        LOG.error("Got unexpected ctrl-c, ignoring")
                     self.packet_io.interrupt_event.clear()
 
                 if self.non_stop and self.is_target_running:
                     try:
                         if self.target.get_state() == Target.State.HALTED:
-                            self.log.debug("state halted")
+                            LOG.debug("state halted")
                             self.is_target_running = False
                             self.send_stop_notification()
                     except Exception as e:
-                        self.log.error("Unexpected exception: %s", e, exc_info=self.session.log_tracebacks)
+                        LOG.error("Unexpected exception: %s", e, exc_info=self.session.log_tracebacks)
 
                 # read command
                 try:
@@ -531,7 +360,7 @@ class GDBServer(threading.Thread):
                             return
 
             except Exception as e:
-                self.log.error("Unexpected exception: %s", e, exc_info=self.session.log_tracebacks)
+                LOG.error("Unexpected exception: %s", e, exc_info=self.session.log_tracebacks)
 
     def handle_message(self, msg):
         try:
@@ -540,7 +369,7 @@ class GDBServer(threading.Thread):
             try:
                 handler, msgStart = self.COMMANDS[msg[1:2]]
             except (KeyError, IndexError):
-                self.log.error("Unknown RSP packet: %s", msg)
+                LOG.error("Unknown RSP packet: %s", msg)
                 return self.create_rsp_packet(b""), 0
 
             if msgStart == 0:
@@ -551,7 +380,7 @@ class GDBServer(threading.Thread):
             return reply, detach
 
         except Exception as e:
-            self.log.error("Unhandled exception in handle_message: %s", e, exc_info=self.session.log_tracebacks)
+            LOG.error("Unhandled exception in handle_message: %s", e, exc_info=self.session.log_tracebacks)
             return self.create_rsp_packet(b"E01"), 0
 
     def extended_remote(self):
@@ -559,12 +388,12 @@ class GDBServer(threading.Thread):
         return self.create_rsp_packet(b"OK")
 
     def detach(self, data):
-        self.log.info("Client detached")
+        LOG.info("Client detached")
         resp = b"OK"
         return self.create_rsp_packet(resp)
 
     def kill(self):
-        self.log.debug("GDB kill")
+        LOG.debug("GDB kill")
         # Keep target halted and leave vector catches if in persistent mode.
         if not self.persist:
             self.board.target.set_vector_catch(Target.VectorCatch.NONE)
@@ -579,7 +408,7 @@ class GDBServer(threading.Thread):
         # handle breakpoint/watchpoint commands
         split = data.split(b'#')[0].split(b',')
         addr = int(split[1], 16)
-        self.log.debug("GDB breakpoint %s%d @ %x" % (data[0:1], int(data[1:2]), addr))
+        LOG.debug("GDB breakpoint %s%d @ %x" % (data[0:1], int(data[1:2]), addr))
 
         # handle software breakpoint Z0/z0
         if data[1:2] == b'0':
@@ -624,7 +453,7 @@ class GDBServer(threading.Thread):
         if not self.is_threading_enabled():
             return self.create_rsp_packet(b'OK')
 
-        self.log.debug("set_thread:%s", data)
+        LOG.debug("set_thread:%s", data)
         op = data[0:1]
         thread_id = int(data[1:-3], 16)
         if not (thread_id in (0, -1) or self.thread_provider.is_valid_thread_id(thread_id)):
@@ -670,7 +499,7 @@ class GDBServer(threading.Thread):
                 self.current_thread_id = currentThread.unique_id
         else:
             if self.current_thread_id != 1:
-                self.log.debug("Current thread %x is no longer valid, switching context to target", self.current_thread_id)
+                LOG.debug("Current thread %x is no longer valid, switching context to target", self.current_thread_id)
                 self.target_facade.set_context(self.target_context)
                 self.current_thread_id = 1
 
@@ -698,7 +527,7 @@ class GDBServer(threading.Thread):
     def resume(self, data):
         addr = self._get_resume_step_addr(data)
         self.target.resume()
-        self.log.debug("target resumed")
+        LOG.debug("target resumed")
 
         if self.first_run_after_reset_or_flash:
             self.first_run_after_reset_or_flash = False
@@ -714,7 +543,7 @@ class GDBServer(threading.Thread):
 
             # Wait for a ctrl-c to be received.
             if self.packet_io.interrupt_event.wait(0.01):
-                self.log.debug("receive CTRL-C")
+                LOG.debug("receive CTRL-C")
                 self.packet_io.interrupt_event.clear()
                 self.target.halt()
                 val = self.get_t_response(forceSignal=signals.SIGINT)
@@ -731,7 +560,7 @@ class GDBServer(threading.Thread):
                             continue
 
                     pc = self.target_context.read_core_register('pc')
-                    self.log.debug("state halted; pc=0x%08x", pc)
+                    LOG.debug("state halted; pc=0x%08x", pc)
                     val = self.get_t_response()
                     break
             except exceptions.Error as e:
@@ -739,7 +568,7 @@ class GDBServer(threading.Thread):
                     self.target.halt()
                 except:
                     pass
-                self.log.warning('Exception while target was running: %s', e, exc_info=self.session.log_tracebacks)
+                LOG.warning('Exception while target was running: %s', e, exc_info=self.session.log_tracebacks)
                 val = ('S%02x' % self.target_facade.get_signal_value()).encode()
                 break
 
@@ -747,7 +576,7 @@ class GDBServer(threading.Thread):
 
     def step(self, data, start=0, end=0):
         addr = self._get_resume_step_addr(data)
-        self.log.debug("GDB step: %s (start=0x%x, end=0x%x)", data, start, end)
+        LOG.debug("GDB step: %s (start=0x%x, end=0x%x)", data, start, end)
         self.target.step(not self.step_into_interrupt, start, end)
         return self.create_rsp_packet(self.get_t_response())
 
@@ -810,7 +639,7 @@ class GDBServer(threading.Thread):
             else:
                 default_action = action
 
-        self.log.debug("thread_actions=%s; default_action=%s", repr(thread_actions), default_action)
+        LOG.debug("thread_actions=%s; default_action=%s", repr(thread_actions), default_action)
 
         # Only the current thread is supported at the moment.
         if thread_actions[currentThread] is None:
@@ -847,18 +676,18 @@ class GDBServer(threading.Thread):
             self.is_target_running = False
             self.send_stop_notification(forceSignal=0)
         else:
-            self.log.error("Unsupported v_cont action '%s'" % thread_actions[1:2])
+            LOG.error("Unsupported v_cont action '%s'" % thread_actions[1:2])
 
     def flash_op(self, data):
         ops = data.split(b':')[0]
-        self.log.debug("flash op: %s", ops)
+        LOG.debug("flash op: %s", ops)
 
         if ops == b'FlashErase':
             return self.create_rsp_packet(b"OK")
 
         elif ops == b'FlashWrite':
             write_addr = int(data.split(b':')[1], 16)
-            self.log.debug("flash write addr: 0x%x", write_addr)
+            LOG.debug("flash write addr: 0x%x", write_addr)
             # search for second ':' (beginning of data encoded in the message)
             second_colon = 0
             idx_begin = 0
@@ -909,10 +738,10 @@ class GDBServer(threading.Thread):
             self.target_context.flush()
             val = hex_encode(bytearray(mem))
         except exceptions.TransferError:
-            self.log.debug("get_memory failed at 0x%x" % addr)
+            LOG.debug("get_memory failed at 0x%x" % addr)
             val = b'E01' #EPERM
         except MemoryAccessError as e:
-            self.log.debug("get_memory failed at 0x%x: %s", addr, str(e))
+            LOG.debug("get_memory failed at 0x%x: %s", addr, str(e))
             val = b'E01' #EPERM
         return self.create_rsp_packet(val)
 
@@ -935,10 +764,10 @@ class GDBServer(threading.Thread):
                 self.target_context.flush()
             resp = b"OK"
         except exceptions.TransferError:
-            self.log.debug("write_memory failed at 0x%x" % addr)
+            LOG.debug("write_memory failed at 0x%x" % addr)
             resp = b'E01' #EPERM
         except MemoryAccessError as e:
-            self.log.debug("get_memory failed at 0x%x: %s", addr, str(e))
+            LOG.debug("get_memory failed at 0x%x: %s", addr, str(e))
             val = b'E01' #EPERM
 
         return self.create_rsp_packet(resp)
@@ -961,10 +790,10 @@ class GDBServer(threading.Thread):
                 self.target_context.flush()
             resp = b"OK"
         except exceptions.TransferError:
-            self.log.debug("write_memory failed at 0x%x" % addr)
+            LOG.debug("write_memory failed at 0x%x" % addr)
             resp = b'E01' #EPERM
         except MemoryAccessError as e:
-            self.log.debug("get_memory failed at 0x%x: %s", addr, str(e))
+            LOG.debug("get_memory failed at 0x%x: %s", addr, str(e))
             val = b'E01' #EPERM
 
         return self.create_rsp_packet(resp)
@@ -987,10 +816,10 @@ class GDBServer(threading.Thread):
 
     def handle_query(self, msg):
         query = msg.split(b':')
-        self.log.debug('GDB received query: %s', query)
+        LOG.debug('GDB received query: %s', query)
 
         if query is None:
-            self.log.error('GDB received query packet malformed')
+            LOG.error('GDB received query packet malformed')
             return None
 
         if query[0] == b'Supported':
@@ -1024,7 +853,7 @@ class GDBServer(threading.Thread):
                 return self.create_rsp_packet(resp)
 
             else:
-                self.log.debug("Unsupported qXfer request: %s:%s:%s:%s", query[1], query[2], query[3], query[4])
+                LOG.debug("Unsupported qXfer request: %s:%s:%s:%s", query[1], query[2], query[3], query[4])
                 return None
 
         elif query[0].startswith(b'C'):
@@ -1064,14 +893,14 @@ class GDBServer(threading.Thread):
 
         for rtosName, rtosClass in RTOS.items():
             try:
-                self.log.info("Attempting to load %s", rtosName)
+                LOG.info("Attempting to load %s", rtosName)
                 rtos = rtosClass(self.target)
                 if rtos.init(symbol_provider):
-                    self.log.info("%s loaded successfully", rtosName)
+                    LOG.info("%s loaded successfully", rtosName)
                     self.thread_provider = rtos
                     break
             except RuntimeError as e:
-                self.log.error("Error during symbol lookup: " + str(e), exc_info=self.session.log_tracebacks)
+                LOG.error("Error during symbol lookup: " + str(e), exc_info=self.session.log_tracebacks)
 
         self.did_init_thread_providers = True
 
@@ -1104,7 +933,7 @@ class GDBServer(threading.Thread):
 
     # TODO rewrite the remote command handler
     def handle_remote_command(self, cmd):
-        self.log.debug('Remote command: %s', cmd)
+        LOG.debug('Remote command: %s', cmd)
 
         safecmd = {
             b'init'  : [b'(ignored for compatibility)', 0],
@@ -1123,7 +952,7 @@ class GDBServer(threading.Thread):
             resp = hex_encode(resp)
         elif cmd.startswith(b'arm semihosting'):
             self.enable_semihosting = b'enable' in cmd
-            self.log.info("Semihosting %s", ('enabled' if self.enable_semihosting else 'disabled'))
+            LOG.info("Semihosting %s", ('enabled' if self.enable_semihosting else 'disabled'))
         elif cmdList[0] == b'set':
             if len(cmdList) < 3:
                 resp = hex_encode(b"Error: invalid set command\n")
@@ -1160,7 +989,7 @@ class GDBServer(threading.Thread):
 
             for cmd_sub in cmdList:
                 if cmd_sub not in safecmd:
-                    self.log.warning("Invalid mon command '%s'", cmd_sub)
+                    LOG.warning("Invalid mon command '%s'", cmd_sub)
                     resp = ('Invalid Command: "%s"\n' % cmd_sub).encode()
                     resp = hex_encode(resp)
                     return self.create_rsp_packet(resp)
@@ -1184,7 +1013,7 @@ class GDBServer(threading.Thread):
 
     def handle_general_set(self, msg):
         feature = msg.split(b'#')[0]
-        self.log.debug("GDB general set: %s", feature)
+        LOG.debug("GDB general set: %s", feature)
 
         if feature == b'StartNoAckMode':
             # Disable acks after the reply and ack.
@@ -1200,7 +1029,7 @@ class GDBServer(threading.Thread):
             return self.create_rsp_packet(b"")
 
     def handle_query_xml(self, query, offset, size):
-        self.log.debug('GDB query %s: offset: %s, size: %s', query, offset, size)
+        LOG.debug('GDB query %s: offset: %s, size: %s', query, offset, size)
         xml = ''
         if query == b'memory_map':
             xml = self.target_facade.get_memory_map_xml()
@@ -1238,7 +1067,7 @@ class GDBServer(threading.Thread):
 
     def syscall(self, op):
         op = to_bytes_safe(op)
-        self.log.debug("GDB server syscall: %s", op)
+        LOG.debug("GDB server syscall: %s", op)
         request = self.create_rsp_packet(b'F' + op)
         self.packet_io.send(request)
 
@@ -1251,7 +1080,7 @@ class GDBServer(threading.Thread):
 
             # Check for file I/O response.
             if packet[0:1] == b'$' and packet[1:2] == b'F':
-                self.log.debug("Syscall: got syscall response " + packet)
+                LOG.debug("Syscall: got syscall response " + packet)
                 args = packet[2:packet.index(b'#')].split(b',')
                 result = int(args[0], base=16)
                 errno = int(args[1], base=16) if len(args) > 1 else 0
@@ -1270,7 +1099,7 @@ class GDBServer(threading.Thread):
 
             if detach:
                 self.detach_event.set()
-                self.log.warning("GDB server received detach request while waiting for file I/O completion")
+                LOG.warning("GDB server received detach request while waiting for file I/O completion")
                 break
 
         return -1, 0
@@ -1291,7 +1120,7 @@ class GDBServer(threading.Thread):
         # Optionally append core
         if self.report_core:
             response += ("core:%x;" % self.core).encode()
-        self.log.debug("Tresponse=%s", response)
+        LOG.debug("Tresponse=%s", response)
         return response
 
     def get_threads_xml(self):
@@ -1333,7 +1162,7 @@ class GDBServer(threading.Thread):
     def event_handler(self, notification):
         if notification.event == Target.Event.POST_RESET:
             # Invalidate threads list if flash is reprogrammed.
-            self.log.debug("Received POST_RESET event")
+            LOG.debug("Received POST_RESET event")
             self.first_run_after_reset_or_flash = True
             if self.thread_provider is not None:
                 self.thread_provider.read_from_target = False
